+++ crates/vertigo-cli/src/serve/mod.rs
@@ -2,9 +2,10 @@ mod html;
 mod mount_path;
 mod request_state;
 mod response_state;
-mod serve_command;
+mod serve_opts;
+mod serve_run;
 mod server_state;
 mod wasm;
 
-pub use serve_command::run;
-pub use serve_command::ServeOpts;
+pub use serve_opts::{ServeOpts, ServeOptsInner};
+pub use serve_run::run;